rename: 에러 타입 파일명 변경
