+++ src/shared/error/error.type.ts
